// frontend/src/model/bus.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Bus Ticket Booking API
 * REST API for searching trips, booking seats and managing bus ticket operations.
 * OpenAPI spec version: v1.0
 */
import type { Operator } from './operator';
import type { BusLayout } from './busLayout';

export interface Bus {
  id?: string;
  operator?: Operator;
  busLayout?: BusLayout;
  plateNumber?: string;
  amenities?: string[];
  isActive?: boolean;
  createdAt?: string;
}