// frontend/src/model/createOperatorRequest.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Bus Ticket Booking API
 * REST API for searching trips, booking seats and managing bus ticket operations.
 * OpenAPI spec version: v1.0
 */
import type { CreateOperatorRequestContactInfo } from './createOperatorRequestContactInfo';

export interface CreateOperatorRequest {
  /** @minLength 1 */
  name: string;
  contactInfo?: CreateOperatorRequestContactInfo;
}