// frontend/src/model/createStationRequest.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Bus Ticket Booking API
 * REST API for searching trips, booking seats and managing bus ticket operations.
 * OpenAPI spec version: v1.0
 */

export interface CreateStationRequest {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  city: string;
  /** @minLength 1 */
  address: string;
}