// frontend/src/model/createTripRequest.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Bus Ticket Booking API
 * REST API for searching trips, booking seats and managing bus ticket operations.
 * OpenAPI spec version: v1.0
 */
import type { PricingRequest } from './pricingRequest';

export interface CreateTripRequest {
  routeId: string;
  busId: string;
  departureTime: string;
  arrivalTime: string;
  pricings: PricingRequest[];
}