// frontend/src/model/operator.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Bus Ticket Booking API
 * REST API for searching trips, booking seats and managing bus ticket operations.
 * OpenAPI spec version: v1.0
 */
import type { OperatorContactInfo } from './operatorContactInfo';

export interface Operator {
  id?: string;
  name?: string;
  contactInfo?: OperatorContactInfo;
  isActive?: boolean;
  createdAt?: string;
}